kr1/requirements.txt
re
pymorphy2
matplotlib
nltk.corpus
nltk.probability
nltk.tokenize